Rename nativeAOT -> appMpowerAot
