+++ frameworks/CSharp/appmpower/src/appMpowerAOT/IJsonSerializer.cs
@@ -1,6 +1,6 @@
 using System.Text.Json;
 
-namespace nativeAOT
+namespace appMpowerAot
 {
    public interface IJsonSerializer<T>
    {
