+++ frameworks/CSharp/appmpower/src/appMpowerAOT/appMpowerAOT.csproj
